requirements.txt
transformers
datasets
pdfminer.six
accelerate
